Move `get_error_message` to the `errors` module
