+++ styleguide_example/api/mixins.py
@@ -4,7 +4,7 @@ from rest_framework import exceptions as rest_exceptions
 from rest_framework.permissions import IsAuthenticated
 from rest_framework_simplejwt.authentication import JWTAuthentication
 
-from styleguide_example.common.utils import get_error_message
+from styleguide_example.api.errors import get_error_message
 
 
 class ApiAuthMixin:
